commands/uninstall: add a --skip-repo option
git lfs install has a --skip-repo option, which prevents any attempt to
modify the repo if run from within one.  Add such an option for git lfs
uninstall as well.  Document it and add tests for its functionality.

+++ commands/command_uninstall.go
@@ -10,7 +10,7 @@ func uninstallCommand(cmd *cobra.Command, args []string) {
 		Error(err.Error())
 	}
 
-	if localInstall || cfg.InRepo() {
+	if !skipRepoInstall && (localInstall || cfg.InRepo()) {
 		uninstallHooksCommand(cmd, args)
 	}
 
@@ -34,6 +34,7 @@ func init() {
 	RegisterCommand("uninstall", uninstallCommand, func(cmd *cobra.Command) {
 		cmd.Flags().BoolVarP(&localInstall, "local", "l", false, "Set the Git LFS config for the local Git repository only.")
 		cmd.Flags().BoolVarP(&systemInstall, "system", "", false, "Set the Git LFS config in system-wide scope.")
+		cmd.Flags().BoolVarP(&skipRepoInstall, "skip-repo", "", false, "Skip repo setup, just uninstall global filters.")
 		cmd.AddCommand(NewCommand("hooks", uninstallHooksCommand))
 	})
 }
